feature: XYNE-54899 removed priority classification
