+++ apps/backend/src/integrations/adapters/social-media/google-play/replySender.ts
@@ -1,5 +1,8 @@
 import { SOCIAL_MEDIA_INTERACTION_TYPES } from '@/integrations/social-media/constants';
-import { BaseInteractionReplySender } from '@/integrations/core/baseInteractionReplySender';
+import {
+  BaseInteractionReplySender,
+  InteractionReplyValidationError,
+} from '@/integrations/core/baseInteractionReplySender';
 import type { InteractionReplyContext } from '@/integrations/core/baseInteractionReplySender';
 import type { NormalizedData } from '@/integrations/core/types';
 import { EmailType } from '@prisma/client';
@@ -8,9 +11,12 @@ import { GOOGLE_PLAY_MAX_REPLY_LENGTH } from './constants';
 import { GOOGLE_PLAY_DEVELOPER_REPLY_SUFFIX } from './transformer';
 
 export class GooglePlayReviewsReplySender extends BaseInteractionReplySender {
-  readonly maxReplyLength = GOOGLE_PLAY_MAX_REPLY_LENGTH;
-
   async sendReply(context: InteractionReplyContext): Promise<NormalizedData> {
+    if (context.body.length > GOOGLE_PLAY_MAX_REPLY_LENGTH) {
+      throw new InteractionReplyValidationError(
+        `Reply cannot exceed ${GOOGLE_PLAY_MAX_REPLY_LENGTH} characters`
+      );
+    }
     const occurredAt = await googlePlayClient.reply(
       context.source.id,
       context.externalThreadId,
